Firmware/requirements.txt
# Radropi Firmware Dependencies
# Core dependencies for asteroid detection system

# Computer Vision and Image Processing
opencv-python==4.8.1.78
numpy==1.24.3

# GPS and Serial Communication
pynmea2==1.19.0
pyserial==3.5

# Raspberry Pi GPIO (only needed on Raspberry Pi)
RPi.GPIO==0.7.1

# Machine Learning and AI (for future model integration)
scipy==1.11.1
scikit-learn==1.3.0

# Data Processing and Logging
pandas==2.0.3

# System and Threading
psutil==5.9.5

# Optional: Deep Learning Frameworks (uncomment if using trained models)
# tensorflow==2.13.0
# torch==2.0.1
# torchvision==0.15.2

# Development and Testing
pytest==7.4.0
pytest-cov==4.1.0

# Documentation
sphinx==7.1.2